Reworking the Parameters file that gets saved
Renamed a bunch of columns. This included needing new key names for several things.

Added Gaussian Smoothing to the list of things that gets saved

Fixed things so that every refresh instance also refreshes the parameters so that Parameters.csv can be accurate

Fixed things so that every Spinbox is bound to a refresh function rather than just some Spinboxes