Ringing Patch
This update mitigates the ringing artifact that occurs when the smoothing window is odd. The artifact occurs when extracting the diagonal because the data is discrete. Prior to this update, every other data point would be the sum of 13 data points and every other data point would be the sum of 14 data point. Now, each is 14, but it oscillates back and forth slightly up the diagonal.